+++ core/src/com/ray3k/skincomposer/desktop/DesktopLauncher.java
@@ -467,7 +467,7 @@ public class DesktopLauncher implements DesktopWorker, Lwjgl3WindowListener {
         config.setResizable(true);
         config.useVsync(true);
         config.setWindowedMode(800, 800);
-        config.setBackBufferConfig(8, 8, 8, 8, 16, 0, 10);
+        config.setBackBufferConfig(8, 8, 8, 8, 16, 2, 10);
         var desktopLauncher = new DesktopLauncher();
         config.setWindowListener(desktopLauncher);
         config.setTitle("Skin Composer - New Project*");