+++ core/src/com/ray3k/skincomposer/dialog/DialogDrawables.java
@@ -30,14 +30,16 @@ import com.badlogic.gdx.files.FileHandle;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.Cursor;
 import com.badlogic.gdx.graphics.Cursor.SystemCursor;
-import com.badlogic.gdx.graphics.PixmapIO.PNG;
 import com.badlogic.gdx.graphics.g2d.TextureAtlas;
 import com.badlogic.gdx.math.MathUtils;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.scenes.scene2d.*;
 import com.badlogic.gdx.scenes.scene2d.actions.Actions;
 import com.badlogic.gdx.scenes.scene2d.ui.*;
-import com.badlogic.gdx.scenes.scene2d.utils.*;
+import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
+import com.badlogic.gdx.scenes.scene2d.utils.Drawable;
+import com.badlogic.gdx.scenes.scene2d.utils.NinePatchDrawable;
+import com.badlogic.gdx.scenes.scene2d.utils.SpriteDrawable;
 import com.badlogic.gdx.utils.Align;
 import com.badlogic.gdx.utils.Array;
 import com.badlogic.gdx.utils.ObjectMap.Values;
@@ -51,6 +53,7 @@ import com.ray3k.skincomposer.UndoableManager.CustomDrawableUndoable;
 import com.ray3k.skincomposer.UndoableManager.DrawableUndoable;
 import com.ray3k.skincomposer.data.*;
 import com.ray3k.skincomposer.data.DrawableData.DrawableType;
+import com.ray3k.skincomposer.dialog.DialogTVG.TvgData;
 import com.ray3k.skincomposer.dialog.DialogTenPatch.TenPatchData;
 import com.ray3k.skincomposer.utils.Utils;
 import com.ray3k.stripe.PopTableClickListener;
@@ -83,6 +86,7 @@ public class DialogDrawables extends Dialog {
     private boolean showingOptions;
     public final static FilterOptions filterOptions = new FilterOptions();
     private FilterInputListener filterInputListener;
+    public boolean showTipTVG;
     
     public interface DialogDrawablesListener {
         void confirmed(DrawableData drawable, DialogDrawables dialog);
@@ -148,12 +152,14 @@ public class DialogDrawables extends Dialog {
                 if (file.isDirectory()) {
                     files.addAll(file.list());
                     iter.remove();
-                } else if (!(file.name().toLowerCase().endsWith(".png") || file.name().toLowerCase().endsWith(".jpg") || file.name().toLowerCase().endsWith(".jpeg") || file.name().toLowerCase().endsWith(".bmp") || file.name().toLowerCase().endsWith(".gif"))) {
+                } else if (!(file.name().toLowerCase().endsWith(".png") || file.name().toLowerCase().endsWith(".jpg")
+                        || file.name().toLowerCase().endsWith(".jpeg") || file.name().toLowerCase().endsWith(".bmp")
+                        || file.name().toLowerCase().endsWith(".gif") || file.name().toLowerCase().endsWith("tvg"))) {
                     iter.remove();
                 }
             }
             
-            var filesLimited = new Array<FileHandle>(files);
+            var filesLimited = new Array<>(files);
             iter = filesLimited.iterator();
             while (iter.hasNext()) {
                 var file = iter.next();
@@ -442,6 +448,10 @@ public class DialogDrawables extends Dialog {
             } else {
                 refreshDrawableDisplayNormal();
             }
+        }
+        if (showTipTVG) {
+            dialogFactory.showTipTVG();
+            showTipTVG = false;
         }
     }
     
@@ -974,6 +984,74 @@ public class DialogDrawables extends Dialog {
                         }
                     });
     
+                    break;
+                case TVG:
+                    //settings
+                    button = new ImageTextButton("Settings", getSkin(), "settings-small");
+                    button.getLabelCell().expandX().left();
+                    root.add(button);
+                    root.row();
+                    button.addListener(hideListener);
+                    button.addListener(handListener);
+                    button.addListener(new ChangeListener() {
+                        @Override
+                        public void changed(ChangeEvent event, Actor actor) {
+                            var drawableData = new DrawableData(drawable);
+                
+                            desktopWorker.removeFilesDroppedListener(filesDroppedListener);
+                            dialogFactory.showDialogTVG(drawable, false, new DialogTVG.DialogTvgListener() {
+                                @Override
+                                public void selected(DrawableData drawableData) {
+                                    drawable.set(drawableData);
+                                    projectData.setChangesSaved(false);
+                                    gatherDrawables();
+                                    atlasData.produceAtlas();
+                                    sortBySelectedMode();
+                                    getStage().setScrollFocus(scrollPane);
+                                    desktopWorker.addFilesDroppedListener(filesDroppedListener);
+                                    refreshDrawableDisplay();
+                                }
+                    
+                                @Override
+                                public void cancelled() {
+                                    desktopWorker.addFilesDroppedListener(filesDroppedListener);
+                                    refreshDrawableDisplay();
+                                }
+                            });
+                        }
+                    });
+        
+                    //duplicate button for an existing ten patch
+                    button = new ImageTextButton("Duplicate", getSkin(), "duplicate-small");
+                    button.getLabelCell().expandX().left();
+                    root.add(button);
+                    root.row();
+                    button.addListener(handListener);
+                    button.addListener(hideListener);
+                    button.addListener(new ChangeListener() {
+                        @Override
+                        public void changed(ChangeEvent event, Actor actor) {
+                            var drawableData = new DrawableData();
+                            drawableData.set(drawable);
+                
+                            dialogFactory.showDuplicateDialog("Duplicate Ten Patch Drawable", "Please enter the name of the duplicated ten patch drawable", drawable.name, new DialogFactory.InputDialogListener() {
+                                @Override
+                                public void confirmed(String text) {
+                                    drawableData.name = text;
+                                    atlasData.getDrawables().add(drawableData);
+                                    gatherDrawables();
+                                    atlasData.produceAtlas();
+                                    sortBySelectedMode();
+                                }
+                    
+                                @Override
+                                public void cancelled() {
+                        
+                                }
+                            });
+                        }
+                    });
+        
                     break;
             }
     
@@ -1647,6 +1725,13 @@ public class DialogDrawables extends Dialog {
                     continue;
                 }
             }
+            
+            if (!filterOptions.tvg) {
+                if (drawable.type == DrawableType.TVG) {
+                    iter.remove();
+                    continue;
+                }
+            }
     
             if (!filterOptions.pixel) {
                 if (drawable.type == DrawableType.PIXEL) {
@@ -1875,7 +1960,7 @@ public class DialogDrawables extends Dialog {
                 }
             }
 
-            List<File> files = desktopWorker.openMultipleDialog("Choose drawable file(s)...", defaultPath, "png,jpg,jpeg,bmp,gif", "Image files");
+            List<File> files = desktopWorker.openMultipleDialog("Choose drawable file(s)...", defaultPath, "png,jpg,jpeg,bmp,gif,tvg", "Image files");
             if (files != null && files.size() > 0) {
                 Gdx.app.postRunnable(() -> {
                     drawablesSelected(files);
@@ -2100,7 +2185,12 @@ public class DialogDrawables extends Dialog {
     private void finalizeDrawables(Array<DrawableData> backup, Array<FileHandle> filesToProcess) {
         for (FileHandle file : filesToProcess) {
             DrawableData data = new DrawableData(file);
-            if (Utils.isNinePatch(file.name())) {
+            if (Utils.isTvg(file.name())) {
+                data.type = DrawableType.TVG;
+                data.file = file;
+                data.tvgData = new TvgData();
+                showTipTVG = true;
+            } else if (Utils.isNinePatch(file.name())) {
                 data.type = DrawableType.NINE_PATCH;
             } else {
                 data.type = DrawableType.TEXTURE;
@@ -2437,6 +2527,7 @@ public class DialogDrawables extends Dialog {
         public boolean tiled = true;
         public boolean custom = true;
         public boolean tenPatch = true;
+        public boolean tvg = true;
         public boolean hidden = false;
         public boolean font = false;
         public boolean regularExpression = false;
@@ -2450,6 +2541,7 @@ public class DialogDrawables extends Dialog {
             tiled = filterOptions.tiled;
             custom = filterOptions.custom;
             tenPatch = filterOptions.tenPatch;
+            tvg = filterOptions.tvg;
             regularExpression = filterOptions.regularExpression;
             name = filterOptions.name;
             hidden = filterOptions.hidden;