+++ core/src/com/ray3k/skincomposer/dialog/DialogExport.java
@@ -33,6 +33,7 @@ import com.badlogic.gdx.scenes.scene2d.ui.*;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.utils.Array;
 import com.ray3k.skincomposer.Main;
+import com.ray3k.skincomposer.data.DrawableData;
 import com.ray3k.skincomposer.data.FontData;
 import com.ray3k.skincomposer.data.FreeTypeFontData;
 import com.ray3k.skincomposer.data.JsonData.ExportFormat;
@@ -110,6 +111,18 @@ public class DialogExport extends Dialog {
             @Override
             public void changed(ChangeListener.ChangeEvent event, Actor actor) {
                 projectData.setExportingFonts(fontCheckBox.isChecked());
+            }
+        });
+    
+        getContentTable().row();
+        var tvgCheckBox = new CheckBox("Copy tvg files to destination", skin);
+        tvgCheckBox.setChecked(projectData.isExportingTVG());
+        getContentTable().add(tvgCheckBox);
+        tvgCheckBox.addListener(handListener);
+        tvgCheckBox.addListener(new ChangeListener() {
+            @Override
+            public void changed(ChangeListener.ChangeEvent event, Actor actor) {
+                projectData.setExportingFonts(tvgCheckBox.isChecked());
             }
         });
         
@@ -235,6 +248,14 @@ public class DialogExport extends Dialog {
                     for (FreeTypeFontData font : projectData.getJsonData().getFreeTypeFonts()) {
                         if (font.useCustomSerializer && !font.file.parent().equals(fileHandle.parent())) {
                             font.file.copyTo(fileHandle.parent());
+                        }
+                    }
+                }
+    
+                if (projectData.isExportingTVG()) {
+                    for (DrawableData drawableData : projectData.getAtlasData().getDrawables()) {
+                        if (!drawableData.file.parent().equals(fileHandle.parent())) {
+                            drawableData.file.copyTo(fileHandle.parent());
                         }
                     }
                 }